lesson8/requirements.txt
requests
pytest
pytest-flake8
python-dotenv
allure-pytest